+++ lib/ButtonModule/include/ButtonModuleInterface.hpp
@@ -66,7 +66,7 @@ public:
      * @param longPressTime The long press time for button triggers.
      * @param timeBetweenDoublePress The time between double presses for button triggers.
      */
-    virtual void startListening(uint32_t usStackDepth = 10000, uint8_t checkInterval = 30, uint8_t debounceTime = 90, uint16_t longPressTime = 1000, uint16_t timeBetweenDoublePress = 500) = 0;
+    virtual void startListening(uint16_t usStackDepth = 10000, uint8_t checkInterval = 30, uint8_t debounceTime = 90, uint16_t longPressTime = 1000, uint16_t timeBetweenDoublePress = 500) = 0;
 
     /**
      * @brief Stops listening for button triggers.
